# GH-129149: Add fast paths to four more `PyLong_From*` functions