--- a/Misc/NEWS.d/next/Core_and_Builtins/2025-03-13-20-23-02.gh-issue-129149.z42wkm.rst
+++ b/Misc/NEWS.d/next/Core_and_Builtins/2025-03-13-20-23-02.gh-issue-129149.z42wkm.rst
@@ -1,0 +1,3 @@
+Add fast path for medium-size integers in :c:func:`PyLong_FromInt32`,
+:c:func:`PyLong_FromUInt32`, :c:func:`PyLong_FromInt64` and
+:c:func:`PyLong_FromUInt64`. Patch by Chris Eibl.