--- a/Objects/longobject.c
+++ b/Objects/longobject.c
@@ -6741,19 +6741,23 @@ PyUnstable_Long_CompactValue(const PyLongObject* op) {
 }
 
 
-PyObject* PyLong_FromInt32(int32_t value) {
+PyObject* PyLong_FromInt32(int32_t value)
+{
     PYLONG_FROM_INT(uint32_t, int32_t, value);
 }
 
-PyObject* PyLong_FromUInt32(uint32_t value) {
+PyObject* PyLong_FromUInt32(uint32_t value)
+{
     PYLONG_FROM_UINT(uint32_t, value);
 }
 
-PyObject* PyLong_FromInt64(int64_t value) {
+PyObject* PyLong_FromInt64(int64_t value)
+{
     PYLONG_FROM_INT(uint64_t, int64_t, value);
 }
 
-PyObject* PyLong_FromUInt64(uint64_t value) {
+PyObject* PyLong_FromUInt64(uint64_t value)
+{
     PYLONG_FROM_UINT(uint64_t, value);
 }