Use the macro `PYLONG_FROM_INT` to implement `PyLong_FromInt32()` and `PyLong_FromInt64()`, and the macro `PYLONG_FROM_UINT` to implement `PyLong_FromUInt32()` and `PyLong_FromUInt64()`, because they have a fast path for small and medium-size integers.

before:
```
PyLong_FromInt32: some small ints
          -5:   228.22 ms
           0:   218.91 ms
         256:   214.52 ms
PyLong_FromInt32: some medium ints
 -1073741823:   313.45 ms
        1000:   196.13 ms
  1073741822:   305.81 ms
PyLong_FromInt32: some bigger ints
 -1073741824:   305.89 ms
  1073741824:   283.47 ms
```
after:
```
PyLong_FromInt32: some small ints
          -5:    21.09 ms
           0:    20.80 ms
         256:    20.27 ms
PyLong_FromInt32: some medium ints
 -1073741823:   100.03 ms
        1000:   101.13 ms
  1073741822:    98.36 ms
PyLong_FromInt32: some bigger ints
 -1073741824:   204.82 ms
  1073741824:   205.22 ms
```

Benchmark done like in https://github.com/python/cpython/pull/129301#issuecomment-2692297879.

<!-- gh-issue-number: gh-129149 -->
* Issue: gh-129149
<!-- /gh-issue-number -->