@chris-eibl 2025-03-13
Pretty much the same as `test_long_as_ssize_t` we've discussed in https://github.com/python/cpython/pull/129301: via round tripping e.g. for in `PyLong_FromUInt32`
https://github.com/python/cpython/blob/1a8e5742cdcf3dba7fc592d036adab49877c42ba/Modules/_testlimitedcapi/long.c#L772-L781
and then using `check_long_asint`, where I've already added more values to cover more cases (especially negative small and medium-size integers): https://github.com/python/cpython/pull/129301#issuecomment-2692176804