@picnixz 2025-03-13
> Maybe add even more values? Definitely no performance bottleneck and might gain test coverage. E.g. one more positive and negative small integer?

If we're already covering the small, medium and large integers, both in negative and positive ranges, it's fine.