@chris-eibl 2025-03-22
@vstinner are you fine with this? (sorry for the ping, though)